refactor: eliminate calls to .node().id() with new trait HasIdentity
