+++ riscv_analysis/src/cfg/mod.rs
@@ -36,6 +36,6 @@ mod interrupt_handler;
 mod ref_cell_replacement;
 pub use ref_cell_replacement::*;
 
-mod node_instruction_properties;
-
 mod node_gen_kill;
+mod node_has_identity;
+mod node_instruction_properties;
