+++ riscv_analysis/src/cfg/node_has_identity.rs
@@ -1,0 +1,9 @@
+use crate::parser::HasIdentity;
+
+use super::CfgNode;
+
+impl HasIdentity for CfgNode {
+    fn id(&self) -> uuid::Uuid {
+        self.node().id()
+    }
+}
