+++ riscv_analysis/src/parser/trait_has_identity.rs
@@ -1,0 +1,6 @@
+use uuid::Uuid;
+
+pub trait HasIdentity {
+    #[must_use]
+    fn id(&self) -> Uuid;
+}
